// resources/js/admin/items/definitions/item-form-state-definition.ts
import { ItemAlchemyType } from '../enums/item-alchemy-type';
import { ItemCatalogType } from '../enums/item-catalog-type';
import { ItemCraftingType } from '../enums/item-crafting-type';
import { ItemDefaultPosition } from '../enums/item-default-position';
import { ItemEffectType } from '../enums/item-effect-type';
import { ItemSkillType } from '../enums/item-skill-type';
import { ItemSpecialtyType } from '../enums/item-specialty-type';

export default interface ItemFormStateDefinition {
  name: string;
  type: ItemCatalogType | '';
  description: string;
  default_position: ItemDefaultPosition | '';
  market_sellable: boolean;
  can_drop: boolean;
  cost: string;
  gold_dust_cost: string;
  shards_cost: string;
  copper_coin_cost: string;
  gold_bars_cost: string;
  alchemy_type: ItemAlchemyType | '';
  specialty_type: ItemSpecialtyType | '';
  base_damage: string;
  base_ac: string;
  base_healing: string;
  base_damage_mod: string;
  base_ac_mod: string;
  base_healing_mod: string;
  str_mod: string;
  dur_mod: string;
  dex_mod: string;
  chr_mod: string;
  int_mod: string;
  agi_mod: string;
  focus_mod: string;
  ambush_chance: string;
  ambush_resistance: string;
  counter_chance: string;
  counter_resistance: string;
  effect: ItemEffectType | '';
  drop_location_id: number | null;
  unlocks_class_id: number | null;
  item_skill_id: number | null;
  skill_name: string;
  skill_bonus: string;
  skill_training_bonus: string;
  fight_time_out_mod_bonus: string;
  move_time_out_mod_bonus: string;
  xp_bonus: string;
  ignores_caps: boolean;
  can_resurrect: boolean;
  resurrection_chance: string;
  spell_evasion: string;
  artifact_annulment: string;
  healing_reduction: string;
  affix_damage_reduction: string;
  devouring_light: string;
  devouring_darkness: string;
  can_craft: boolean;
  craft_only: boolean;
  crafting_type: ItemCraftingType | '';
  skill_level_required: string;
  skill_level_trivial: string;
  usable: boolean;
  can_stack: boolean;
  lasts_for: string;
  stat_increase: boolean;
  increase_stat_by: string;
  damages_kingdoms: boolean;
  kingdom_damage: string;
  affects_skill_type: ItemSkillType | null;
  increase_skill_bonus_by: string;
  increase_skill_training_bonus_by: string;
  can_use_on_other_items: boolean;
  holy_level: string;
  gains_additional_level: boolean;
}