// resources/js/admin/items/api/definitions/item-request-definition.ts
import { ItemAlchemyType } from '../../enums/item-alchemy-type';
import { ItemCatalogType } from '../../enums/item-catalog-type';
import { ItemCraftingType } from '../../enums/item-crafting-type';
import { ItemDefaultPosition } from '../../enums/item-default-position';
import { ItemEffectType } from '../../enums/item-effect-type';
import { ItemSkillType } from '../../enums/item-skill-type';
import { ItemSpecialtyType } from '../../enums/item-specialty-type';

export default interface ItemRequestDefinition {
  name: string;
  type: ItemCatalogType;
  description: string;
  default_position: ItemDefaultPosition | null;
  market_sellable: boolean;
  can_drop: boolean;
  cost: number | null;
  gold_dust_cost: number | null;
  shards_cost: number | null;
  copper_coin_cost: number | null;
  gold_bars_cost: number | null;
  alchemy_type: ItemAlchemyType | null;
  specialty_type: ItemSpecialtyType | null;
  base_damage: number | null;
  base_ac: number | null;
  base_healing: number | null;
  base_damage_mod: number | null;
  base_ac_mod: number | null;
  base_healing_mod: number | null;
  str_mod: number | null;
  dur_mod: number | null;
  dex_mod: number | null;
  chr_mod: number | null;
  int_mod: number | null;
  agi_mod: number | null;
  focus_mod: number | null;
  ambush_chance: number | null;
  ambush_resistance: number | null;
  counter_chance: number | null;
  counter_resistance: number | null;
  effect: ItemEffectType | null;
  drop_location_id: number | null;
  unlocks_class_id: number | null;
  item_skill_id: number | null;
  skill_name: string | null;
  skill_bonus: number | null;
  skill_training_bonus: number | null;
  fight_time_out_mod_bonus: number | null;
  move_time_out_mod_bonus: number | null;
  xp_bonus: number | null;
  ignores_caps: boolean;
  can_resurrect: boolean;
  resurrection_chance: number | null;
  spell_evasion: number | null;
  artifact_annulment: number | null;
  healing_reduction: number | null;
  affix_damage_reduction: number | null;
  devouring_light: number | null;
  devouring_darkness: number | null;
  can_craft: boolean;
  craft_only: boolean;
  crafting_type: ItemCraftingType | null;
  skill_level_required: number | null;
  skill_level_trivial: number | null;
  usable: boolean;
  can_stack: boolean;
  lasts_for: number | null;
  stat_increase: boolean;
  increase_stat_by: number | null;
  damages_kingdoms: boolean;
  kingdom_damage: number | null;
  affects_skill_type: ItemSkillType | null;
  increase_skill_bonus_by: number | null;
  increase_skill_training_bonus_by: number | null;
  can_use_on_other_items: boolean;
  holy_level: number | null;
  gains_additional_level: boolean;
}